Implement dimensional stiffness/damping for TrajectoryTaskGeneric

+++ src/mc_tasks/RelativeEndEffectorTask.cpp
@@ -89,7 +89,7 @@ void configure_pos_task(std::shared_ptr<mc_tasks::PositionTask> & t,
   }
   if(config.has("positionStiffness"))
   {
-    t->stiffness(config("positionStiffness"));
+    t->stiffness(static_cast<double>(config("positionStiffness")));
   }
 }
 
@@ -108,7 +108,7 @@ void configure_ori_task(std::shared_ptr<mc_tasks::OrientationTask> & t,
   }
   if(config.has("orientationStiffness"))
   {
-    t->stiffness(config("orientationStiffness"));
+    t->stiffness(static_cast<double>(config("orientationStiffness")));
   }
 }
